Move functions to mesh_rouintes.
Moved:

* _create_trimesh
* FreeBoundary class
* get_freeboundaries
