+++ openep/draw/__init__.py
@@ -1,7 +1,6 @@
 __all__ = ['draw_routines']
 
 from .draw_routines import (
-    get_freeboundaries,
     draw_free_boundaries,
     draw_map,
 )
